Remove 'exam' parameter from the ResponseNode.example method
Having the exam parameter added another level of hierarchy for example definitions which was unnecessary.  This made it impossible to give a response example that returned an array without wrapping it within a hash object. Index calls need to return an array. Now things like the below will be possible

swagger_path '/items' do
  operation :get do
    response 200 do
      key :description, 'OK'
      example do
        key :"application/json", [item1_json, item2_json]
     end
  end
end